backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
google-auth
requests
flet-map
